move carrier info mut. to toolbox.go
