+++ graphql2/graphqlapp/mutation.go
@@ -2,15 +2,12 @@ package graphqlapp
 
 import (
 	context "context"
-	"net/url"
 
 	"github.com/target/goalert/assignment"
 	"github.com/target/goalert/graphql2"
-	"github.com/target/goalert/notification/twilio"
 	"github.com/target/goalert/permission"
 	"github.com/target/goalert/user"
 	"github.com/target/goalert/validation"
-	"github.com/target/goalert/validation/validate"
 
 	"github.com/pkg/errors"
 )
@@ -31,38 +28,6 @@ func (a *Mutation) SetFavorite(ctx context.Context, input graphql2.SetFavoriteIn
 		return false, err
 	}
 	return true, nil
-}
-
-type safeErr struct{ error }
-
-func (safeErr) ClientError() bool { return true }
-
-func (a *Mutation) DebugSendSms(ctx context.Context, input graphql2.DebugSendSMSInput) (*graphql2.DebugSendSMSInfo, error) {
-	err := permission.LimitCheckAny(ctx, permission.Admin)
-	if err != nil {
-		return nil, err
-	}
-
-	err = validate.Many(
-		validate.Phone("To", input.To),
-		validate.Phone("From", input.From),
-		validate.Text("Body", input.Body, 1, 1000),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	msg, err := a.Twilio.SendSMS(ctx, input.To, input.Body, &twilio.SMSOptions{
-		FromNumber: input.From,
-	})
-	if err != nil {
-		return nil, safeErr{error: err}
-	}
-
-	return &graphql2.DebugSendSMSInfo{
-		ID:          msg.SID,
-		ProviderURL: "https://www.twilio.com/console/sms/logs/" + url.PathEscape(msg.SID),
-	}, nil
 }
 
 func (a *Mutation) TestContactMethod(ctx context.Context, id string) (bool, error) {
