+++ graphql2/graphqlapp/toolbox.go
@@ -6,14 +6,49 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 
 	"github.com/target/goalert/config"
 	"github.com/target/goalert/graphql2"
+	"github.com/target/goalert/notification/twilio"
 	"github.com/target/goalert/permission"
+	"github.com/target/goalert/validation/validate"
 	"github.com/ttacon/libphonenumber"
 )
 
 const twLookupURL = "https://lookups.twilio.com/v1/PhoneNumbers/"
+
+type safeErr struct{ error }
+
+func (safeErr) ClientError() bool { return true }
+
+func (a *Mutation) DebugSendSms(ctx context.Context, input graphql2.DebugSendSMSInput) (*graphql2.DebugSendSMSInfo, error) {
+	err := permission.LimitCheckAny(ctx, permission.Admin)
+	if err != nil {
+		return nil, err
+	}
+
+	err = validate.Many(
+		validate.Phone("To", input.To),
+		validate.Phone("From", input.From),
+		validate.Text("Body", input.Body, 1, 1000),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	msg, err := a.Twilio.SendSMS(ctx, input.To, input.Body, &twilio.SMSOptions{
+		FromNumber: input.From,
+	})
+	if err != nil {
+		return nil, safeErr{error: err}
+	}
+
+	return &graphql2.DebugSendSMSInfo{
+		ID:          msg.SID,
+		ProviderURL: "https://www.twilio.com/console/sms/logs/" + url.PathEscape(msg.SID),
+	}, nil
+}
 
 func (a *Mutation) DebugCarrierInfo(ctx context.Context, input graphql2.DebugCarrierInfoInput) (*graphql2.DebugCarrierInfo, error) {
 	// must be admin to fetch carrier info
